Reuse notification title as its id
The change allows to reuse one notification for the same conversation.
When a notification id is reused, send_notification will replace the
old notification and show it again as if it was new.

The notification title usually is the name of the conversation, which
is a valid indicator for us.