Provide an application menu
The resources are arranged to take advantage of the Gtk.Application automatic
resource loading. This allows getting them loaded with minimum code changes:
the only thing needed is to register a Gio.Resouce for the built resources
file.

The items from the application menu are non-functional, but they get loaded
and displayed in the top-bar menu under GNOME Shell, or as the window-icon
menu in the application window header bar under other DEs.