Track active notifications and withdraw them when focused
As discussed in issue #3, while we cannot track which notfications correspond
to which rooms in order to withdraw only the notifications caused by the room
being displayed, it is better than nothing to withdraw all the notifications.
That happens to be what many IM applications do anyway (i.e. WhatsApp on
Android), so it won't startle users.