Tweak allowed zoom range and preferences slider
A maximum zoom level of 2.0 is already quite a lot, so use that as maximum
value. An interface bigger than than that would be ridiculous. Also, do not
draw the zoom value, as it has a range around 1.0 which may not be meaningful
to users (a different story would be if the value would show up as a
percentage).