Load StatusIcon menu model and icons from resources
This uses the automatic "<appid>/icons/" location added by Gtk.Application to
the default Gtk.IconTheme. Note that a 16x16@2x group of icons is also added,
to be used in HiDPI displays when the 16px icons are requested; these are
copies  of the 32px icons.

For the menu, a model is included in the "<appid>/gtk/menus.ui" resource,
which also gets automatically loaded by Gtk.Application. This avoids having to
manually construct the menu items.