Make MainWindow a proper subclass of Gtk.ApplicationWindow
Apart from making things more natural, some code bits are also simplified, and
the headerbar spinner is now bound to the "network-busy" property, effectively
decoupling it. Ideally the status icon can be decoupled now from the webview
by connecting it to the "notify::network-busy" signal, but that is a refactor
to do another time.

Fixes #18.