Expose a setting to toglle the WebKitGTK+ inspector
This adds a new "enable-developers-tools" setting, which for now controls
whether access to the built-in WebKitGTK+ inspector is enabled. The name
of the setting is generic enough to allow using it as a toggle for other
development oriented features in the future.

Additionally, this also adds a switch to the preferences window which allows
toggling the setting.

Fixes #20.