Add hidden setting to disable header bars
This is particularly useful for tiling window managers, or just to maximize
vertical screen space usage. As the people likely to want to toggle this would
be power users, it is fine to keep it as a hidden option. It can be toggled
with:

  gsettings set org.perezdecastro.Revolt:/org/perezdecastro/Revolt/ use-header-bar true

(This is a bit convoluted because the settings schema is relocatable, which we
need to support custom Flatpak builds, but it gets the job done. Of course
using dconf-editor is a tad more convenient.)