Status icon prefixes shouldn't depend on the application id
These are fixed and copied over to the theme directories, and they will
be always the same regardless of which application id is being used at
runtime.

This should fix the bug with the status icons from issue #58 and there
are good chances that this will help #87 and #83 as well.